Rename: 세부페이지 관련 폴더 생성 후 파일 이동
