+++ src/components/post/CompletedComments.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function CompletedComments() {
+  return <div>CompletedComments</div>;
+}
+
+export default CompletedComments;
