+++ src/components/post/CompletedNotification.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function CompletedNotification() {
+  return <div>CompletedNotification</div>;
+}
+
+export default CompletedNotification;
